Comfy now builds and installs, Fixed issues with Mac Signing
Comfy now installs and then packaged into electron 

fixed issues with code signing, there were too many files and the runner was choking on it. There is a PR request to fix this issue on osx-sign but has not been merged. For now use that function as a shim and place it inside of node_modules.